Fixed Authentication Error
