+++ database.js
@@ -1,7 +1,5 @@
 // Importing Firebase and its services
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-app.js";
-import { getAnalytics } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-analytics.js";
-import { getAuth, createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-auth.js";
 import { getDatabase, ref, set } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-database.js";
 
 const firebaseConfig = {
@@ -17,18 +15,16 @@ const firebaseConfig = {
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-const auth = getAuth(app);
 const database = getDatabase(app);
 
 // Function to write apartment data to the database
-function writeApartmentData(apartmentId, bedroom, bathroom, price, zipcode) {
+function writeApartmentData(apartmentId, Bedroom, Bathroom, Price, ZipCode) {
     const reference = ref(database, 'apartments/' + apartmentId);
     set(reference, {
-        Bedroom: bedroom,
-        Bathroom: bathroom,
-        Price: price,
-        ZipCode: zipcode
+        Bedroom: Bedroom,
+        Bathroom: Bathroom,
+        Price: Price,
+        ZipCode: ZipCode
     })
     .then(() => {
         alert("Data saved successfully!");
@@ -38,33 +34,9 @@ function writeApartmentData(apartmentId, bedroom, bathroom, price, zipcode) {
     });
 }
 
-// Event listener for the register submit button
-document.getElementById('authForm').addEventListener("submit", function (event) {
-    event.preventDefault();
-    console.log('Auth form submitted');
-
-    const email = document.getElementById('email').value;
-    const password = document.getElementById('password').value;
-
-    createUserWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        const user = userCredential.user;
-        console.log('User created: ', user);
-        alert("Account created successfully!");
-        window.location.href = "favorites.html";
-      })
-      .catch((error) => {
-        const errorCode = error.code;
-        const errorMessage = error.message;
-        console.error('Error during account creation: ', errorMessage);
-        alert(errorMessage);
-      });
-});
-
 // Event listener for the save apartment button
 document.getElementById('apartmentForm').addEventListener("submit", function (event) {
     event.preventDefault();
-    console.log('Apartment form submitted');
 
     const apartmentId = document.getElementById('apartmentId').value;
     const bedroom = document.getElementById('bedroom').value;
